cleaned and refactored
